+++ packages/account-eip6963/src/emitter.ts
@@ -1,9 +1,14 @@
-import { SophonIcon, type SophonNetworkType } from '@sophon-labs/account-core';
+import {
+  AccountServerURL,
+  SophonIcon,
+  type SophonNetworkType,
+} from '@sophon-labs/account-core';
 import { announceEip6963Provider } from './eip6963';
 import { createSophonEIP1193Provider } from './provider';
 
 export function createSophonEIP6963Emitter(
   network: SophonNetworkType = 'testnet',
+  authServerUrl: string = AccountServerURL[network],
 ) {
   // Skip on server-side
   if (typeof window === 'undefined') {
@@ -11,7 +16,7 @@ export function createSophonEIP6963Emitter(
   }
 
   // Create the provider
-  const provider = createSophonEIP1193Provider(network);
+  const provider = createSophonEIP1193Provider(network, authServerUrl);
 
   // Announce it via EIP-6963
   announceEip6963Provider({
@@ -24,5 +29,7 @@ export function createSophonEIP6963Emitter(
     provider,
   });
 
-  console.log(`Sophon EIP-6963 provider announced for ${network}`);
+  console.log(
+    `Sophon EIP-6963 provider announced for ${network} and url ${authServerUrl}`,
+  );
 }