+++ packages/account-eip6963/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sophon-labs/account-eip6963
 
+## 2.0.0-next.35
+
+### Patch Changes
+
+- Provide authServerUrl to eip6963
+- Updated dependencies
+  - @sophon-labs/account-core@2.0.0-next.35
+
 ## 2.0.0-next.34
 
 ### Patch Changes